+++ cxx/include/enhanced_enum/enhanced_enum.hh
@@ -6,6 +6,8 @@
 #ifndef ENHANCED_ENUM_HH_INCLUDED_
 #define ENHANCED_ENUM_HH_INCLUDED_
 
+#include "details/ranges.hh"
+
 #include <array>
 #include <optional>
 #include <type_traits>
@@ -70,16 +72,6 @@ struct enum_base {
         return static_cast<std::ptrdiff_t>(size());
     }
 
-    /** \brief Get range over all enumerators
-     *
-     * \return A random access range containing all enumerators in the
-     * order they are declared in the enum
-     */
-    static constexpr decltype(auto) all() noexcept
-    {
-        return *(&_enums_array);
-    }
-
     /** \brief Get iterator to the first enumerator
      *
      * \return A random access iterator to the beginning of the range
@@ -88,7 +80,8 @@ struct enum_base {
      */
     static constexpr auto begin() noexcept
     {
-        return _enums_array.begin();
+        return details::enum_iterator {
+            EnhancedEnum {static_cast<LabelEnum>(0)}};
     }
 
     /** \brief Get iterator one past the last enumerator
@@ -97,7 +90,18 @@ struct enum_base {
      */
     static constexpr auto end() noexcept
     {
-        return _enums_array.end();
+        return details::enum_iterator {
+            EnhancedEnum {static_cast<LabelEnum>(size())}};
+    }
+
+    /** \brief Get range over all enumerators
+     *
+     * \return A range containing all enumerators in the order they
+     * are declared in the enum
+     */
+    static constexpr auto all() noexcept
+    {
+        return details::enum_range {begin(), end()};
     }
 
     /** \brief Get the enumerator associated with \p value
@@ -165,19 +169,6 @@ struct enum_base {
     }
 
 private:
-
-    static constexpr auto _init_enums_array() noexcept
-    {
-        constexpr auto N = size();
-        auto ret = std::array<EnhancedEnum, N> {};
-        for (auto i = 0u; i < N; ++i) {
-            ret[i] = static_cast<LabelEnum>(i);
-        }
-        return ret;
-    }
-
-    constexpr static auto _enums_array = _init_enums_array();
-
     LabelEnum label;
 };